Freetype long to unsigned int issue

+++ Header Files/FreetypeRender.h
@@ -22,7 +22,8 @@ class FreetypeRender
 			unsigned int TextureID;
 			glm::ivec2 Size;
 			glm::ivec2 Bearing;
-			unsigned int Advance;
+			//unsigned int Advance;
+			long Advance;
 		};
 		unsigned int VAO, VBO;
 		std::unordered_map<char, Character> Characters;
